refactor(utils): make more generic functions for fetching blockchain events (#33)
* refactor(utils): make more generic functions for fetching blockchain events

* commit the rest
