+++ scripts/calculateRevenue/protocols/types.ts
@@ -1,0 +1,4 @@
+export interface BlockTimestampData {
+  height: number
+  timestamp: number
+}
